+++ packages/fx-core/src/common/templatesUtils.ts
