+++ packages/fx-core/src/plugins/resource/function/ops/deploy.ts
@@ -7,6 +7,9 @@ import AdmZip from "adm-zip";
 import axios from "axios";
 import ignore, { Ignore } from "ignore";
 
+import { WebAppsListPublishingCredentialsResponse } from "@azure/arm-appservice/esm/models";
+import { PluginContext } from '@microsoft/teamsfx-api';
+
 import { AzureInfo, CommonConstants, DefaultValues, FunctionPluginPathInfo } from "../constants";
 import {
   ConfigFunctionAppError,
@@ -21,7 +24,6 @@ import { ErrorMessages, InfoMessages } from "../resources/message";
 import { FunctionEvent, FunctionLanguage } from "../enums";
 import { LanguageStrategyFactory } from "../language-strategy";
 import { Logger } from "../utils/logger";
-import { WebAppsListPublishingCredentialsResponse } from "@azure/arm-appservice/esm/models";
 import { execute } from "../utils/execute";
 import { forEachFileAndDir } from "../utils/dir-walk";
 import { BackendExtensionsInstaller } from "../utils/depsChecker/backendExtensionsInstall";
@@ -29,9 +31,8 @@ import { DotnetChecker } from "../utils/depsChecker/dotnetChecker";
 import { FuncPluginAdapter } from "../utils/depsChecker/funcPluginAdapter";
 import { funcPluginLogger } from "../utils/depsChecker/funcPluginLogger";
 import { FuncPluginTelemetry } from "../utils/depsChecker/funcPluginTelemetry";
-import { PluginContext } from "@microsoft/teamsfx-api";
 import { TelemetryHelper } from "../utils/telemetry-helper";
-import { sendRequestWithRetry } from "../../../../common/templates-helper";
+import { sendRequestWithRetry } from "../../../../common/templatesUtils";
 
 export class FunctionDeploy {
   public static async getLastDeploymentTime(componentPath: string): Promise<Date> {
